requirements.txt
click==8.3.2
einops==0.8.2
fastapi==0.135.3
fsspec==2026.3.0
huggingface_hub==0.36.2
langchain==1.2.15
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.6
numpy==2.2.6
openai==2.31.0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
psutil==7.2.2
pydantic==2.12.5
PyYAML==6.0.3
requests==2.33.1
safetensors==0.7.0
sentencepiece==0.2.1
sse-starlette==3.3.4
streamlit==1.56.0
torch==2.10.0
torchaudio==2.10.0
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
uvicorn==0.44.0
vllm==0.19.0
